+++ mqttwarn/services/pushsafer.py
@@ -36,6 +36,9 @@ def pushsafer(**kwargs):
     if not kwargs['k']:
         kwargs['k'] = os.environ['PUSHSAFER_TOKEN']
 
+    # Don't submit empty parameters to Pushsafer.
+    filter_empty_parameters(kwargs)
+
     url = urllib.parse.urljoin(PUSHSAFER_API, "api")
     data = urllib.parse.urlencode(kwargs).encode('utf-8')
     req = urllib.request.Request(url, data)
@@ -45,6 +48,16 @@ def pushsafer(**kwargs):
 
     if data['status'] != 1:
         raise PushsaferError(output)
+
+
+def filter_empty_parameters(params: t.Dict[str, str]):
+    """
+    Filter empty parameters.
+    """
+    for key in list(params.keys()):
+        value = params[key]
+        if value == "":
+            del params[key]
 
 
 def plugin(srv, item):
@@ -132,7 +145,12 @@ class PushsaferParameterEncoder:
             params['s'] = addrs[3]
 
         if len(addrs) > 4:
-            params['v'] = addrs[4]
+            # Special handling for `vibration`.
+            # With the v1 configuration layout, it should not always trigger a vibration when left empty.
+            # Therefore, a synthetic value `0` can be used to configure "device-default" vibration.
+            value = str(addrs[4])
+            if value != "":
+                params['v'] = value
 
         if len(addrs) > 5:
             params['u'] = addrs[5]