+++ tests/services/pushsafer/test_pushsafer_v2.py
@@ -106,7 +106,8 @@ variants = [
     IoTestItem(id="device-id", in_addrs={"private_key": TEST_TOKEN, "device": "52|65|78"}, out_data={"d": "52|65|78"}),
     IoTestItem(id="icon", in_addrs={"private_key": TEST_TOKEN, "icon": "test.ico"}, out_data={"i": "test.ico"}),
     IoTestItem(id="sound", in_addrs={"private_key": TEST_TOKEN, "sound": "test.mp3"}, out_data={"s": "test.mp3"}),
-    IoTestItem(id="vibration", in_addrs={"private_key": TEST_TOKEN, "vibration": 1}, out_data={"v": "1"}),
+    IoTestItem(id="vibration-default", in_addrs={"private_key": TEST_TOKEN, "vibration": 0}, out_data={"v": ""}),
+    IoTestItem(id="vibration-two", in_addrs={"private_key": TEST_TOKEN, "vibration": 2}, out_data={"v": "2"}),
     IoTestItem(
         id="url",
         in_addrs={"private_key": TEST_TOKEN, "url": "http://example.org"},